[UPGRADE]upgraded lwm modules
