--- lwm/.settings/org.eclipse.core.resources.prefs
@@ -1,7 +1,0 @@
-eclipse.preferences.version=1
-encoding//controllers/__init__.py=utf-8
-encoding//controllers/controllers.py=utf-8
-encoding//models/__init__.py=utf-8
-encoding//models/stock.py=utf-8
-encoding/__init__.py=utf-8
-encoding/__manifest__.py=utf-8
