+++ lwm/__init__.py
@@ -1,4 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from . import controllers
 from . import models
