--- lwm/data/data.xml
@@ -1,6 +1,0 @@
-<odoo>
-    <data noupdate="1">
-    
-    	
-    </data>
-</odoo>
