+++ lwm/models/__init__.py
@@ -1,3 +1,9 @@
 # -*- coding: utf-8 -*-
 
+from . import account
+from . import product
+from . import purchase
+from . import res_company
+from . import res_config
+from . import sale
 from . import stock
