--- lwm/security/lwm_security.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-<data>
-
-
-</data>
-</odoo>
