+++ lwm/views/res_config_views.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+    <data>
+
+        <record id="lwm_res_config_settings_view_form" model="ir.ui.view">
+            <field name="name">lwm.res.config.settings.view.form.inherit.account</field>
+            <field name="model">res.config.settings</field>
+            <field name="priority" eval="40"/>
+            <field name="inherit_id" ref="account.res_config_settings_view_form"/>
+            <field name="arch" type="xml">
+                <xpath expr="//div[@data-key='account']" position="inside">
+                    <h2>Discount</h2>
+                        <div class="row mt16 o_settings_container">
+                            <div class="col-xs-12 col-md-6 o_setting_box" title="This account is for all discounts allocated.">
+                                <div class="o_setting_left_pane"/>
+                                <div class="o_setting_right_pane">
+                                    <!-- <label string='Default Discount Account' for="discount_account_id"/> -->
+                                    <span class="o_form_label">Default Discount Account</span>
+                                    <span class="fa fa-lg fa-building-o" title="Values set here are company-specific." groups="base.group_multi_company"/>
+                                    <div class="text-muted">
+                                        Default discount account to be used for all discounts
+                                    </div>
+                                    <div class="content-group">
+                                        <div class="row mt16">
+                                            <label string="Discount Account" for="discount_account_id" class="col-md-3 o_light_label"/>
+                                            <field name="discount_account_id"/>
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                </xpath>	         
+            </field>
+        </record>
+
+    </data>
+</odoo>
