+++ lwm/views/sale_views.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+    <data>
+
+        <record id="lwm_sale_order_form_view" model="ir.ui.view">
+            <field name="name">lwm.sale.order.form</field>
+            <field name="model">sale.order</field>
+            <field name="inherit_id" ref="sale.view_order_form"/>
+            <field name="arch" type="xml">
+                <field name="tax_totals_json" position="after">
+                    <div class="oe_subtotal_footer_separator oe_inline o_td_label">
+                        <label for="sub_total" />
+                    </div>
+                    <field name="sub_total" nolabel="1" class="oe_subtotal_footer_separator" widget='monetary' options="{'currency_field': 'currency_id'}"/>
+                    <field name="discounted_amount" widget='monetary' options="{'currency_field': 'currency_id'}"/>
+                </field>	
+            </field>
+        </record>
+    
+        <record id="lwm_view_quotation_tree" model="ir.ui.view">
+            <field name="name">lwm.sale.order.tree</field>
+            <field name="model">sale.order</field>
+            <field name="inherit_id" ref="sale.view_quotation_tree"/>
+            <field name="arch" type="xml">
+                <field name="amount_total" position="before">
+                    <field name="total_ordered_quantity"/>
+                </field>	
+            </field>
+        </record>
+
+    </data>
+</odoo>
