+++ lwm/views/stock_views.xml
@@ -1,88 +1,7 @@
 <odoo>
   <data>
-  
-    <record id="lwm_sale_order_form_view" model="ir.ui.view">
-        <field name="name">lwm.sale.order.form</field>
-        <field name="model">sale.order</field>
-        <field name="inherit_id" ref="sale.view_order_form"/>
-        <field name="arch" type="xml">
-	        <field name="amount_tax" position="after">
-	        	<div class="oe_subtotal_footer_separator oe_inline o_td_label">
-                	<label for="sub_total" />
-                </div>
-                <field name="sub_total" nolabel="1" class="oe_subtotal_footer_separator" widget='monetary' options="{'currency_field': 'currency_id'}"/>
-                <field name="discounted_amount" widget='monetary' options="{'currency_field': 'currency_id'}"/>
-	        </field>	
-        </field>
-    </record>
-    
-    <record id="lwm_view_quotation_tree" model="ir.ui.view">
-        <field name="name">lwm.sale.order.tree</field>
-        <field name="model">sale.order</field>
-        <field name="inherit_id" ref="sale.view_quotation_tree"/>
-        <field name="arch" type="xml">
-	        <field name="amount_total" position="before">
-                <field name="total_ordered_quantity"/>
-	        </field>	
-        </field>
-    </record>
 
-  	<record id="lwm_invoice_form" model="ir.ui.view">
-        <field name="name">lwm.account.invoice</field>
-        <field name="model">account.invoice</field>
-        <field name="inherit_id" ref="account.invoice_form"/>
-        <field name="arch" type="xml">
-	        <field name="amount_tax" position="after">
-	        	<div class="oe_subtotal_footer_separator oe_inline o_td_label">
-                	<label for="sub_total" />
-                </div>
-                <field name="sub_total" nolabel="1" class="oe_subtotal_footer_separator" widget='monetary' options="{'currency_field': 'currency_id'}"/>
-                <field name="discounted_amount" widget='monetary' options="{'currency_field': 'currency_id'}"/>
-	        </field>
-        </field>
-    </record>
     
-    <record id="lwm_invoice_tree" model="ir.ui.view">
-        <field name="name">lwm.account.invoice.tree</field>
-        <field name="model">account.invoice</field>
-        <field name="inherit_id" ref="account.invoice_tree"/>
-        <field name="arch" type="xml">
-	        <field name="origin" position="after">
-                <field name="total_ordered_quantity"/>
-	        </field>
-        </field>
-    </record>
-    
-    
-    <record id="lwm_res_config_settings_view_form" model="ir.ui.view">
-            <field name="name">lwm.res.config.settings.view.form.inherit.account</field>
-            <field name="model">res.config.settings</field>
-            <field name="priority" eval="40"/>
-            <field name="inherit_id" ref="account.res_config_settings_view_form"/>
-            <field name="arch" type="xml">
-	            <xpath expr="//div[@data-key='account_invoicing']" position="inside">
-	            	<h2>Discount</h2>
-                        <div class="row mt16 o_settings_container">
-                            <div class="col-xs-12 col-md-6 o_setting_box" title="This account is for all discounts allocated.">
-                                <div class="o_setting_left_pane"/>
-                                <div class="o_setting_right_pane">
-                                    <label string='Default Discount Account'/>
-                                    <span class="fa fa-lg fa-building-o" title="Values set here are company-specific." groups="base.group_multi_company"/>
-                                    <div class="text-muted">
-                                        Default discount account to be used for all discounts
-                                    </div>
-                                    <div class="content-group">
-                                        <div class="row mt16">
-                                            <label string="Discount Account" for="discount_account_id" class="col-md-3 o_light_label"/>
-                                            <field name="discount_account_id"/>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-	            </xpath>	         
-           </field>
-   	</record>
-   	
+
   </data>
 </odoo>
